wip: fix issue where the globe icon sometimes wouldn't appear
