+++ apps/app/src/electron/triggersHandler.ts
@@ -19,9 +19,9 @@ import {
 	prepareTriggersAreaMap,
 } from '../lib/triggers/keyDisplay/keyDisplay'
 import { SessionHandler } from './sessionHandler'
-import { convertSorensenToElectron } from '../lib/util'
 import { globalShortcut } from 'electron'
 import EventEmitter from 'events'
+import { convertSorensenToElectron } from '../lib/triggers/identifiers'
 
 export interface TriggersHandlerEvents {
 	failedGlobalTriggers: (identifiers: Readonly<Set<string>>) => void
