+++ apps/app/src/lib/util.ts
@@ -23,7 +23,6 @@ import { assertNever, deepClone } from '@shared/lib'
 import shortUUID from 'short-uuid'
 import _ from 'lodash'
 import { describeTimelineObject } from './TimelineObj'
-import { AnyTrigger } from '../models/rundown/Trigger'
 
 export const findGroup = (rundown: Rundown, groupId: string): Group | undefined => {
 	return rundown.groups.find((g) => g.id === groupId)
@@ -792,37 +791,4 @@ export function unReplaceUndefined(obj: any): any {
 		return newObj
 	}
 	return obj
-}
-
-/**
- * Translate a trigger's identifier keys from the Sorensen format into the Electron format.
- * The Electron format does not distinguish between left and right modifier keys.
- * See https://www.electronjs.org/docs/latest/api/accelerator for more information.
- */
-export function convertSorensenToElectron(identifier: string): string {
-	const cleanedIdentifier = identifier.replace(/^keyboard-(?:Key)?/, '')
-	switch (cleanedIdentifier) {
-		case 'ShiftLeft':
-		case 'ShiftRight':
-			return 'Shift'
-		case 'ControlLeft':
-		case 'ControlRight':
-			return 'Control'
-		case 'MetaLeft':
-		case 'MetaRight':
-			return 'Meta'
-		case 'NumpadEnter':
-			return 'Enter'
-		case 'AltLeft':
-		case 'AltRight':
-			return 'Alt'
-		case 'OSLeft':
-		case 'OSRight':
-			return process.platform === 'darwin' ? 'CmdOrCtrl' : 'Super'
-		default:
-			return cleanedIdentifier
-	}
-}
-export function triggerIsKeyboard(trigger: AnyTrigger, index: number): boolean {
-	return trigger.fullIdentifiers[index]?.startsWith('keyboard')
 }
